refactor: move tsx components to relevant folders
